eemon_551/requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
blinker==1.6.2
cffi==1.15.1
click==8.1.7
comm==0.1.3
contourpy==1.0.6
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
docker
dj-database-url==2.1.0
Django
django-environ==0.11.2
executing==1.2.0
filelock
Flask
fonttools
greenlet==2.0.2
gunicorn==21.2.0
ipykernel
ipython
itsdangerous
jedi==0.18.2
Jinja2
jupyter_client
jupyter_core
kiwisolver
MarkupSafe
matplotlib
matplotlib-inline
mpmath
mysqlclient
nest-asyncio
numpy
opencv-contrib-python
opencv-python
packaging
parso
pexpect
pickleshare
Pillow
platformdirs
prompt-toolkit
psutil
ptyprocess
pure-eval
pycparser
Pygments
PyMySQL
Pillow
pyparsing
python-dateutil
pyzmq
psycopg2
requests
scipy
six
SQLAlchemy
sqlparse
stack-data
sympy
tornado
traitlets
typing_extensions
virtualenv
wcwidth
Werkzeug
whitenoise
pip==23.3.1
djangorestframework==3.12.4

